👌 IMP: Move TranspotitionTable to transposition_table.rs
